Moved OnboardingActivity.kt to the view directory
