+++ app/src/main/AndroidManifest.xml
@@ -49,7 +49,7 @@
             android:screenOrientation="portrait" />
 
         <activity
-            android:name=".view.onboarding.OnboardingActivity"
+            android:name=".view.OnboardingActivity"
             android:screenOrientation="portrait" />
 
         <activity
