+++ app/src/main/java/org/hackillinois/android/view/MainActivity.kt
@@ -24,7 +24,6 @@ import org.hackillinois.android.common.JWTUtilities
 import org.hackillinois.android.notifications.FirebaseTokenManager
 import org.hackillinois.android.view.home.HomeFragment
 import org.hackillinois.android.view.leaderboard.LeaderboardFragment
-import org.hackillinois.android.view.onboarding.OnboardingActivity
 import org.hackillinois.android.view.profile.ProfileFragment
 import org.hackillinois.android.view.scanner.ScannerFragment
 import org.hackillinois.android.view.scanner.StaffScannerFragment
